Added padding and margins, added accordions
used HTML details and summary tags to create accordion as an answer to the challenge to not use JS (although I wanted to practice JS in this challenge. Maybe next time).
styled the bottom border of the details containers as dividers rather than hr

removed the default details arrow and created inline-box spans to separate the summary text from the supplied arrow image.

Next up: make the summary text bold and arrow point.up when details is open.