questions now change to bold and arrows flip when details is clicked!
I am ECSTATIC that the arrow flip now works PERFECTLY!!! I tried so may things. First, I had to remove the default arrows from the details tag. I accomplished that by following the steps here: https://stackoverflow.com/questions/56758098/how-to-position-detail-marker-to-come-after-summary Then I separated the questions from the arrows with spans of the appropriate width. Then I aligned the arrows to the right. Then it was a puzzle getting the arrows to flip when details clicked.

Some things I tried/thought about doing that did not work
-my first idea was to use :after as explained in the above-referenced article, but I knew that wouldn't work, because I was only supplied a down-pointing arrow in the challenge.
-then I spent a lot of time trying to use lots of css pseudo classes. :target actually worked, but only while the left mouse button was depressed. Then it would flip right back.
-then I tried the method described in this documentation: https://developer.mozilla.org/en-US/docs/Web/CSS/:target-within , by making the questions into undecorated links and the arrows the targets with ids, but that totally did not work at all, and it stopped the details from opening.
-Next, I tried adding the transform: rotate to the details open property and had a break-through. It flipped the entire summary, including the question and the arrow!
-I tried to add .arrow after summary in the details [open] property, but that caused everything to just shut down, and the color and font-weight would not change.
-Finally, I made a separate details[open] property with .arrow and transform: rotate isolated and IT TOTALLY WORKS!!!!!!!!!!!!!